Refactor k-tag
